🎨 Restarted the project with a new mindset.
The previous version was a little overly designed, therefore,
this version aims to be very agile, with decently modularized components,
but does not over-emphasize writing perfect code and forcing good design,
as it creates too much pressure for myself, and slowing the progress to a crawl.

To quote the Zen of Python,
"Special cases aren't special enough to break the rules.
Although practicality beats purity.
Now is better than never.
Although never is often better than *right* now."

As I have reached a stage where I do not need to consciously write code for it to be good,
I will embrace this mentality and try to write good enough code,
which will, given time to refactor, be better and better and become perfect eventually.