🚩 Setup after porting over.
🛠️ Fix mypy install types.

Mypy requires installation of different types in order to work. It was not included in the workflow file.

📝 Updated documentations and moved some issues in the old repo over.

The issues are now tracked inside improvement proposals, where it makes sense.

♻️ Renamed `Column` to `Buffer`.

The name `Column` is too overloaded, so I would like to move away from it.

I also updated the documentation in the improvement proposals.

🚨 Removed cached table because it might not be the best abstraction.

A better abstraction might be `.compute()` + `.reset()` per each iteration,
s.t. there are ways to know whether this iteration has passed.

🔌 Made changes to `Block` to expose features of tensordicts.

Making the `Block` trainable is a good idea, as it's only used in the torch backend.

👩‍❤️‍👨 Type hints and `Relation` visitor class improvement.

`Relation`'s visitor class is pulled out into a standalone class,
to shorten typing and improvement of type hints because mypy isn't that advanced yet.

🚚 Refactored s.t. all `.Visitor` nested class are now top level classes.

This helps static checkers in verifying types when generics are involved.

🔧 Tests fix.

Tests were previously broken due to visitor's `__call__` changes.

🏞️ `Tracer`'s outline is done.

I'll have to add in models later at some stage.

🦇 Version bump and miscellaneous update.

I'm updating the auto ml part of the system, and studying related libraries.

⏰ Unfinished commit on metadata for tensors.

The meatadata is very important during the compilation phase to provide additional constraints.

💱 Choices added for previews.

This is used in `Preview`'s parameter tuning.

🧪 Tests for `Interval`s added.

Currently, there are 4 kinds of intervals, int float bool str,
all of which is tested in this commit.

🔭 `Preview` added.

`Preview`s are responsible for the metadata of a model / node in processing,
such that we know everything about a model except runtime behavior before initialization.

Since `Preview` is a `Node[Preview]` type,
it should be able to adjust the costs / parameters / etc on the fly,
for example, a `LazyLinear` should be able to change its parameters according to the inputs.

📛 Changed `Interval` to `Range`.

The name better captures the output of a function, whereas intervals are befitting for values that exist on a number line.