🎹 Typing changes and added `Preview`.
🌇 `Preview`s are now separate from `Info`s.

This change is made to further differentiate between runtime information and what the backend should provide.
`Preview` is someting that the backend should provide, during construction of which no runtime info is available.
`Info` is closer to my project `koila`'s metadata and stores the "pre-passes" of runtime information,
so long as input nodes are available.

😑 A simple, greedy compiler is added.

This would be the first step towards implementing optimizing compilers, but at this time, let's focus on getting the greedy compiler to work.

♻️ Broken down `logics` module and adopted a flatter structure.

`logics.{relations,exprs}` -> `relalg`, `logics.trees` -> `plans`, `logics.dtypes` -> `dtypes`.

📊 Renamed `dtypes` to `schemas` as well as improvements.

The name `schemas` is more fitting for the module and works better thematically, considering that the naming of the APIs should follow numpy / torch as much as possible.

🟰 Improved `__eq__` function of the `DataType`.

Enhanced the function to be able to compare with strings and self.

🔡 Conversion utilities for `ArrayDtype`s.

Allow all datatypes to be promoted to the most generic `ArrayDtype` would help in doing arithmetic operations.

🪢 Intermediate nodes used by `Relation` made internal.

The (now) `RelNode` class is no longer exported from the module because it is just an intermediate contract of the `Relation` nodes.

🚃 `Preview`s are moved to their own module.

This should help reduce clutter as well as making each module easier to understand.