🚣 Another set of refactor changes.
🛳️ Moved `Einsum` into `previews`.

This is because I am going to work on it, as well as the model part, and I think it fits better over there.

🌋 Streamlined `TableSchema` in preparation for integration with `Einsum`.

In particular, I would like creation of `TableSchema` to be very fast, as well as index accessing, such that the table schemas can be easily compared and created on the fly.

🧪 Fixed tests for `TableSchema`s.

This was missing in the previous commit.

⚠️ Custom error code such that errors are tracable.

This should help both me and other users in the future to understand what's going on.

⛄ Testing if `networkx` type stubs are correctly recognized.

Somehow it wasn't installed by `mypy --install-types`, so I'm not sure if it'll work. Let's push and see.

🐛 Bug is filed to `mypy` (python/mypy#18365).

Since `networkx`'s type stubs are not added, to pass CI, I ignored the typing of `networkx` for now.

🚚 Moved `Dag` to `plans`.

DAGs should be implementation agnostic to be able to be used by multiple downstream tasks.

🐶 `Dag`s are added and tested.

Removed `networkx` dependency because it's overkill and has `typing` issues (python/mypy#18365).

🤿 Typing fix.

Github action should pass now.

🪛 Fixed some smaller issues.

Typing, imports, moved to using `aioway`'s internal error system.

⚠️ Some more errors added.

Those are the "core" errors used by many different files.

👨‍🔧 Fix format.

Prevent the formatting workflow from failing.

✍️ Updated some more design constraints.

I have decided that `Block` would be purely `torch`,
unless in the future I notice that new functionality can be added in an implementation agnostic manner.