💑 Started to work on relational algebra operators.
💪 `Buffer`s and `Record`s are ok.

Now, let's work on `Block`s.

🚷 `Block` is done.

Made `Block` a concrete type for now, simplifying the design.

☠️ Removed redundant methods from `Block`s.

Previously the methods are not used and are duplicate of existing methods.

🔧 Fix type issues.

Temporarily renamed the imports from `TensorBlock` to `Block`.

🧱 `Block` is made abstract (again).

This is final. I have decided that `Block` vs `Stream` are implementations, and `Table`s are trackers.

🏷️ `TableSchema` renamed to `Schema`.

As it is somewhat of a public API, the name `Schema` is better.

🚧 Continually working on abstracting `Block` and `Buffer`.

They are now split into several subclasses, as well as having a factory.

🧲 Pulling common functionality like `join`, `concat` out.

This is to provide some shared functionality and allow for strategy pattern on different ways of computation on `Block`s.

⚗️ Filter operations added.

This resides in the blocks module currently.

⏲️ Small improvements to the `Exec` implementations.

I have decided to use the `Block` abstraction and simplify designs.

🥖 `Batch` added.

Need to test it but it should be the core data structure in memory.

🪓 Split `blocks` into `blocks`, `buffers`, `batches`.

It is split this way because they represent different abstractions, and that the resources are handled separately (even tho e.g. torch block can only make use of torch buffers).