🪟 Added `Frame` and `Stream` and integrate `Block`.
🖼️ `Frame` is ok.

`Frame` is a bounded data structure on top of `Block`. It would be responsible for providing relational algebra constructs by combining `Exec`s once those are ready.

🎏 Added `Stream` and made improvements to `Block` slicing.

Added some basic implementations of `Stream` and made some changes in other places accordingly.

🚢 Added `BlockFrame`, replacing both `TensorFrame` and `PandasFrame`.

This is to allow code reuse and avoid duplicate logic.

🛟 `Frame` and `Block` now returns `Buffer` instead of `NDArray`.

This makes it more and more framework agnostic and elegant.

🆎 Module `types` renamed to `typing`.

This is because model `types` is better fitted to describe model's types.

🪧 Added `buffer` factory function, prepare for testing.

Updated dependencies also.

🛻 Renamed the backend modules to their module names.

For example, `TensorBlock` is renamed to `TorchBlock` to be consistent with the conversion method names.