✂️ Split `schemas` and `attrs`. (#11)
* ✂️ Split `schemas` and `attrs`.

This is due to the fact that they are different concepts fundamentally.

* 🎈 Update `Attributes.to` function.

As `Attributes` becomes a core type, this added functionality is needed.

🗼 Design docs for `Caster`.

Still considering to remove the multiple implentations and add them back later.
But for now, let's just wing it and see if it is too much to handle.