💑 Relational algebra operators and refactors.
🚧 Work in progress for some physical operators.

In this commit, filters are first implemented and thought over.

⚕️ Work in progress for `Frame`s and `Stream`s.

I'm going to re-introduce `Block` as a wrapper around `TensorDict` to provide simple checks. See: #9, #12.

🛒 Grouping `Frame`s, `Stream`s into `Table`s. Reintroduce `Block` (fix #16).

I have decided to introduce a `Table` that acts like `RDD`s in `spark`.

Fix: #16
Fix: #13

💑 Some relational algebra operators added.

I'm in the process of integrating some relational algebra operators, but still need to check for schema.

⏪ Unmerge `frames` and `streams`.

This job should be reserved for #19.

➗ Added more relation algebra frame operators.

Removed some types that are causing issues.

🕹️ Simplified and merged relations into `TableSchema` (#15).

This simplification is done s.t. frame implementation can be easier.

🖼️ Relational operators for `Frame`s added (fix #10).

Added some relational operators for `Frame`s, which are bounded tables.

Fix: #10
Fix: #13
Fix: #15
Fix: #16

🧪 Testing of `Block` in progress.

I'll finish it ASAP.

🧱 `Block` is now tested.

Next I'll work on `Frame`s.

ℹ️ Formatted the output.

This shall make github actions pass.

🔣 Added ``sympy`` support in column expressions.

This should save us a lot of implementing ``Expr``s ourselves, as we can directly leverage ``sympy``'s symbol processing.

〽️ `BlockFrame` tested.

Just needed to test the rest of the `Frame`s and we could merge the PR.

1️⃣ Unary frame operator testing done.

Next, binary frame operator.

🧪 Tests for projection frame.

I will now move on to binary frames.

🔀 Merge and concat frames tested.

This PR is ready to merge.

👤 Singleton factory added (fix #24).

I have written a super basic factory, that is essentially just a dict, but supports singleton.

🌄 Added `typing.override` for some `abc.abstractmethod`s (fix #20).

This shall help `mypy` to uncover bugs (in the future, now it doesn't seem to do anything).

🧑‍🔬 Documented physical designs (fix #31).

In this commit, I have outlined the scope of each components,
why each is desirable and used in what context.

This is the documentation for #22.

The missing part is to better generalize `Block`,
such that it works with libraries like `torch_geometric`.

🪇 Unary frames have been ported to streams (#26).

Binary frames remaining.

Note on `block_frame_iter` as it created some really subtle bugs due to iterator behavior,
which made this small refactor take a long time:

> Do not use this iterator in tests for comparison,
> because the iterator is instantiated once for each tests.
> Let's look at an example:
> >>> to_test = PassThroughStream(iter)
> >>> golden = iter
> >>> for a, b in zip(to_test, golden):
> >>>    # This would fail!
> >>>    assert a is not b
> Both sides of the equation uses the same iterator underneath,
> and it would create subtle bugs.

💥 Delete moved files.

This wasn't included in the previous commit somehow.

🎏 Done moving from frames to streams.

In this commit, binary frames are moved to streams.

However, in join frames, the RHS is still required to be a frame.

🧪 Use fixtures in tests (fix #25).

This improves efficiency slightly with fixtures' scope control.

💹 Updated depdencies into groups (fix #34).

In the future, optional depdencies should stay optional.

💐 Miscellaneous improvements and re-batching frame.

Changes include a new re-batching `Frame` type,
and several different magic function addition.

I'll be working on `Index` next: #27.