🪚 Changes realted to executors.
🥷 Grouped executions into `execs`.

Tables, frames, streams, and indices are grouped.

✂️ The names are cut short in documentation.

To keep inline with the code.

📃 Moved docs to the modules implemented (#43).

This is not yet done, nor is it possible to do at this time, because not all designs are implemented.

🔢 Version bump for dependencies.

Now using the latest torch and tensordict.

🏭 `Factory` is added.

This is very useful for registering classes and manage subclasses easily with `__init_subclass__`.

💥 Removed `Factory` (#24).

I have decided against implementing factories because using dictionaries is easier.

🪓 Split `Stream` and `Exec` (fix #44).

I also did a refactor, not yet finalized but will move `streams` and `frames` to `tabular`.

🔃 Allow `DataLoader` to be used in `Exec` (fix #45).

Next, port `Frame` and `Stream` to be `Dataset`s to leverage this change.