💾 Some data related changes (frames, streams, tensordict).
🚛 Moved frames and streams together.

This makes more sense logically.

🧪 Moved tests to a new location.

For the tests to match the code.

🐞 Right now `TensorDict` seems to hang at some functions.

Not sure why deadlocks happen. I would like to see if this is the case on github actions as well.

🧑‍🤝‍🧑 Join is working now and decently performant.

Fixed some performance issue.

🎨 Templates for indices added.

This is useful for organizing indices by their backends.

📑 Added `TensorDictFrame`.

For now it's still heavily reliant on `Block`.