📺 Rendering and display the internal nodes.
🌳 `TreeNode` and rendering added (#39).

I'll have to test it, but how?

🗂️ Moving `indices` to under `frames`.

Indexes only work on bounded datasets, so this move makes sense.
In addition, it's possible that there are more efficient e.g. hierarchical representation s.t. indexing is fast / not needed.

🗑️ Remove unused file.

Fix the github actions issue.

➕ Added another `Exec` specifically to track `Frame`s and `Stream`s.

This is useful for explain query down the road.

🧑‍🏫 Explain is OK (close: #39).

Use `reprlib` or `pprint` and do not rely on `rich` in the future.

ℹ️ Ran formatter.

Fix github actions.