👴 Added einsum related features.
🧑‍🏫 `Einsum` added.

I believe `np.einsum` has both easy to use interface and are powerful, making them the ideal candidate to use in attrset matching.

💬 Representation for `Einsum` added.

This means that they should be rendered properly now.

🎒 Refactored some files, removed genre.

Einsum feels much more natural and powerful to handle. Might be easier to index too.

🤹 `Einsum` can handle multiple parameters now in its RHS.

This is necessary to allow modelling multi-return in models / tables.

🚧 WIP for symbolic attrs passing / processing.

This shall be accomplished with einsum, as well as a couple other utilities.

⏮️ Added pre-commit hooks.

Moved my manual formatting into the pre-commit hooks.

🎭 Added genres as a way to constrain the different

🆙 Updated dependencies.

This is for pre-commits.

☀️ `Einsum` attributes working.

Now I'll have to implement genres and functions s.t. constraining is easy.

🐦 Migrate to `pre-commit` in github actions, and update poetry.

This should save some typing down the road.

🧐 Type checking moved to pre-commit.yaml.

This is just so all pre commit checks can be in one place.

👶 Pre-commit workflow renamed.

This goes more in line with other files.

🖥️ Allow `Einsum` types to be computable.

This makes implementation of `AttrSetFunc` easier.

♟️ `AttrSetFunc` backed by `Einsum` is now done.

I can move on to implement some losses now.

🚧 Split `Einsum` and `EinsumAttrFunc`.

This makes it easier to maintain in the future, rather than a parser that would parse and construct, now parsing and verifying happen in different stages.

🛠️ Fixed the types of the `EinsumAttrSetFunc`, and some refactor.

Refactored functions into `funcs`.