🥷 More executor changes.
☎️ Changed `Exec` to use a `__call__` interface over `__next__`.

This would make shared computation a lot easier.

🚀 Callstack for execution implemented.

This would help keeping track of the execution flow of the program and possibly store / cache computation.

🚧 WIP for execution contexts.

Once I am able to interact with the execution call easily, this is done.
This would make it much easier to manage the running algorithm.

📛 Possibly no aliases for `Exec`?

Think about how to achieve that.

📅 Contexts are moved to plans.

I also refined a little how contexts are to be used, which should be implicit and invisible.

For `Exec`, where I originally introduced contexts, I'll probably use some `__init_subclass__` magic to rewrite every subclass's `execute` to use contexts, this way the functions would still appear to be simple recursive function rather than a double recursion, at the same time not needing to handle duplicate calls in 1 exececution pull.

⏭️ Using `next` instead of `execute`.

This is a revert of the previous code. With context we simply need to define a recursive function, rather than a doubly recursive function.