🧑‍💻 Processor changes.
💻 Added a couple processor implementation.

Still WIP.

🌯 Wrapping procs and sequential procs added.

These container procs are used to wrap other procs and are useful for composing complex managers.

🛟 Lifetime manager for processors.

This is useful in cases where you want to ensure that the processor is initialized, but some contexts need to be different per every call.

🍫 Tests done for wrappers.

Now test other processors.

📝 Memoization of processors tested.

Next I would need to test the remaining processors.

🖼️ `FrameExec` revamped to leak less implementation details.

In this commit I also removed the `logs` module and all its tests, due to having no use.

🧪 Tests for stack added.

I'll need to test scope, and that should be all.

💬 Moved some comments.

This is just so I can first work on something else in another branch.

🗺️ Executor for modules added.

Pulled the map executor out and added a new executor for modules.

🌹 Pretty printing for trees.

I have set up a display class to do exactly that. Right now the API is a little ugly but I'll revisit later.

🛟 Working on lifetime manager.

Once that is done, I can release v0.0.2.

🪼 Lifetime component is tested.

I just need a little final touch before release.

😎 Simplified processor interface.

Now, processor is no longer a function to function, but rather a function wrapper.
Rewriter is a processor to processor.

📞 Using `Callable` directly instead of procs.

This should make the code more flexibile.

📝 Changed `MemoCall` to using `functools`.

Moved the implementation to using standard library mechanism, for more robustness.

🩼 Helper function for `OpaqueCall`.

This is used for quick substitution of opaque pointer.

🪵 Bring back `logs` (for lazy logging).

Turns out that this is still needed for logging `torch.Tensor` unless we want to do a full subclass.