🔭 Misc changes (inspection of states, refactors).
👣 Tracer for executors.

This makes constructing executors way easier.

🔦 Torch modules added to planning test.

Still pending execution tests.

📟 Calls back to procs.

This makes implementing factories way easier.

🏭 Factory for processor added and tested.

This makes it possible to create a new instance without coding it.

🫠 Tidied up the modules module.

I would like to work on execution dags first, so I'll leave this for tomorrow.

⚓ Broken down execs for better modularity.

This is to make way for the future optimizers

🤜 Version bump to 0.0.3 in pyproject.

This is to prepare for the next release.

🐛 Fix `DataLoader` unable to serialize default value.

I'm using a global function instead of a lambda function to avoid serialization issues.

✏️ Made some refactors and made todo comments into tickets.

This should improve the tracking of the tickets and make it easier to find them in the long run.