🤏 Minor changes and renames.
🐛 Bug fix for attribute set.

Fixed a typo.

🎥 Added prototypes for codecs (#48).

This would allow working with IO heavy (compressed) binary formats and non tensor types.

🔁 Repeat executor added.

This is useful for memoization purposes, where we would need to cache the same exact block for use later.

👪 Moved children into tree nodes.

This would solve some duplication code problems.

🚚 Renamed `TreeNode` to `Node`.

The nodes we use are not purely trees, so renaming is in order.

🔊 `EchoExec` added.

This executor uses a local state to track and repeat the input. It is crucial for memoization.

🧪 Added additional checks for `EchoExec`.

Ensure that the inputs are properly echoed, and does not terminate prematurely.