🧾 Adopting more relational operators to the new iterable system.
🤝 Operator `MATCH` is added.

This is useful as the atomic operation for both hash-join and merge-join.

🪆 Nested loop's operation ported over as operators.

🚠 Moved more implementations over to ops / poller system.

I also made some small changes to address the removal of impls module in indices.

🚛 Completly moved the implementations over to pollers + ops.

Now, I'll have to decide a syntax for constructing the new abstraction.

🎫 Keys of `Poller` and `Op` now have `_n` for `n`-ary operation.

This helps avoiding confusion, and the `n`-ary suffix is checked universally in the base class (`Nargs`).

🏭 Preset factory is done.

Finally mapped the exec's registry to the new pollre + op.

🔦 Remove tests for torch for now.

These shall be re-added after torch modules are implemented.

✅ Added compile / init time checks for `ARGC` as well as runtime ones.

🪚 Refined initialization method of `Exec`.

Now `Exec` acts both as a delegate constructor that tranparently passes the arguments to poller and op.

❓ Pending fix of `Poller` for `Frame`.

Ideally `FrameOp`'s functionality should live in `Poller`, but it's currently not. Let's try to do that.