🐞 Some bug fixing changes.
🚧 Done writing the `Op` interfaces. Now porting over tests.

🪢 WIP on the reduction of operators.

Right now there is a bug in the reduction making the answer incorrect.

✏️ `Op`s are working now.

It is working now, but the usage is a little awkard, because they are considered generator function as of now.
This doesn't work well with the `BlockIter` interface which also wraps the dependency.

⛑️ Safe defaults: always clone the iterator.

The PR should be ready to merge.

📝 Added some documentation for the `Op` operators.

👷 Only use the default `num_workers` in `DataLoader` (fix #88).

This is because we are operating on `TensorDict` which uses in-memory array format,
so multiprocessing only slows it down.

In this commit I also refined somewhat the fixtures in `ops` tests.

ℹ️ Formatted a misbehaving file.

🧫 Renamed convenient function.

This is because I would like to change it to return a `Thunk` object, which `Exec` would operate on.