🛠️ Fix some previous bugs.
♟️ DAG is ready and tested.

`DagExec` uses `LazyExec` as a backbone, with `RepeatOp` inserted.
This part can be refactored into the `Compiler` in the future.

🫙 Fix previous commit not inserting `RepeatOp` properly (fix #95).

In the previous commit the `Thunk`s are not modified properly,
s.t. the tree transformation is not complete,
and node list in DAG would still reference the old `Thunk`.

This is caused by `Thunk`s being immutable,
therefore we need to modify the entire list everytime there is a rewrite.

This part should make into the `Compiler`, as previous commit suggests.

🛠️ Fix typing issue in github actions.

Might need some test improvements in the future.

💬 Fixed grammar mistake in docs.

🤏 Small change to `Thunk`.

⛔ Xfail the DAG executor, currently.

This will be fixed later, after refactor of `Block`.