🔖 Changes related to in-memory batches.
🚧 Working on making a new `Block` that implements a subset of `TensorDict`.

This would allow us to track runtime behavior with it, while only testing on `TensorDict`.

WIP. Mypy checks would fail.

🧱 Added `Block` back (fix #102).

This is also related to #91, due to `Block` being added back,
so it needs to be mostly compatible with `TensorDict`,
as well as being easy to construct.

🎞️ `Frame` refinement is done.

I have successfully merged `PandasFrame` and `TorchFrame`.

🔙 Reverted the `Block` addition (fix #104).

🤏 Small refactor.