🔎 Inspect the internal states of execution with contexts.
🏄 Contexts for iterator based exec (fix #90).

In this PR I would like to ensure that contexts are working properly, and can be used to derive useful information like some tracing.

💫 Passing contexts transparantly in iteration. This is a WIP.

This way eliminates the shared state `ctx`, and ensures that the contexts can be easily swapped.

🐞 Still buggy implementation.

I'm committing s.t. I can decouple execs and contexts first, before fixing the bugs.

🚛 Further refactored context module.

Now I can focus on debugging why the contexts doesn't always get applied correctly.

🪓 Split the `ExecStack` into its own file.

Added some icecream debugging utilities.

💥 Removed registries for now.

I can add that back later, but it's not too pressing.

💈 Remove caching in `pdm` setup.

As the lock file still exists, it should still be fast anyways.

🤡 Forgot to remove all caching.

This should fix github actions.

☎️ Caller is implemented, but need some tuning.

I think using a `ExecCtxDict` is a good idea, making it very flexible.

🛕 Temporarily remove caller context.

This is because since an `Exec` can have multiple callers, this suddenly becomes not that useful.

🍳 Simplify the contexts part.

As it doesn't do much (because `Exec`'s implementation cannot depend on it),
I have decided to simplify the part. In the case where it will be needed in the future,
such as distrbuted / network case, this would not be the abstraction to use,
as the `Exec` itself would need to be made aware of it.

📛 Using `ctxl` for the abbrieviation of `contextlib`.

There are multiple other constructs already named `ctx`.