🫟 Misc changes on ops and attrs.
⚛️ Decided that `Op` is atomic.

💰 First commit for cost computation.

I would like to refactor the costs into their own kernels.

✂️ Dissect `thunks` from `ops`.

`Thunk`s should be able to hold both `Thunk`s and `Exec`s.
This is why they do not belong to `ops`.

🚛 Deprecate the `aioway.attrs` package.

It's a little too difficult to use,
as it aims to parse information quite high level,
making writing low level `kernels` a lot of overhead.

I am removing it now, and make it s.t. it's easy to use in the `kernels` package,
and then we can expose those for high level usage.