🎏 `Stream` interface.
`Stream` is an external iterator, this is because previously we used `Generator`,
and it is quite buggy simply because `Generator`s are good at iterating, but difficult to inspect the states.
The external iterator design would also make tracking computation (slightly) easier and clearer.
This makes debugging easier, allowing us to check not just the content but also the iteration.

🗺️ Migrate the old `Plan`s to `Stream`s.

I have decided to use `TensorDict` as the `Stream` input and output,
this might change in the future depending on how kernel functions and tracing are implemented.

Tests still fail in this commit.

🧪 Implemented and fix tests in this commit.

I have both migrated all the old tests for `Plan`s and added some additional ones,
to check on the iteration of the `Stream`s.

Linting with `pylint` is enforced on `aioway.streams` package (only) for now.

🤲 `Stream`s are now using `TableStream`.

Also, `Stream`s are streamlined,
their `__len__` attribute is defined to be the length of the entire stream.

🧼 Re-organize `Table`s and added tests for them.

Now `Table`s are integrated in a better manner to `Stream`s,
where they can convert to one another.

☣️ Deleted `execs` module.

No longer needed.

📍 Pin version on 3.13 for now.

Due to `tensordict` not supporting 3.14 well now, I'll pin it to just 3.13.